+++ CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGELOG for FlatCAM beta
 - fixed a small typo when creating a new Gerber empty App Object
 - in Copper Thieving Plugin updated the GUI
 - in Extract Drills Plugin updated the GUI
+- in Paint Plugin updated the GUI
 
 4.09.2021